+++ src/shared/modules/tabs/components/NavTab.vue
@@ -5,6 +5,7 @@
  * @module shared/modules/tabs/NavTab
  * @vue-prop {String} id - the id of the navTab-button.
  * @vue-prop {String} label - the label used on the navTab-button.
+ * @vue-prop {String} icon - optional bootstrap icon class suffix (e.g. 'bi-geo-alt') shown before the label.
  * @vue-prop {Boolean} active - whether the tab is currently active.
  * @vue-prop {String} target - used to specify the id of the element shown by the navTab button (i.e. '#section-1').
  * @vue-prop {Function} interaction - can be used to bind a function to an interaction with the navTab-button, to be executed on click.
@@ -19,6 +20,11 @@ export default {
         label: {
             type: String,
             required: true
+        },
+        icon: {
+            type: String,
+            required: false,
+            default: ""
         },
         active: {
             type: Boolean,
@@ -57,6 +63,12 @@ export default {
             :aria-label="label"
             @click="interaction"
         >
+            <i
+                v-if="icon"
+                class="bi me-2"
+                :class="icon"
+                aria-hidden="true"
+            />
             {{ $t(label) }}
         </button>
     </li>